KOGITO-6130 Review containers used in tests
- Disable DevServices from Quarkus where these are not used
- Review containers versions to avoid duplicates in the build pipeline
