+++ addons/common/persistence/jdbc/src/test/java/org/kie/persistence/jdbc/AbstractProcessInstancesIT.java
@@ -20,11 +20,13 @@ import java.util.Optional;
 
 import javax.sql.DataSource;
 
-import org.junit.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
+import org.drools.core.io.impl.ClassPathResource;
+import org.junit.jupiter.api.Test;
+import org.kie.kogito.auth.IdentityProviders;
+import org.kie.kogito.auth.SecurityPolicy;
 import org.kie.kogito.persistence.jdbc.JDBCProcessInstances;
 import org.kie.kogito.process.ProcessInstance;
+import org.kie.kogito.process.ProcessInstanceReadMode;
 import org.kie.kogito.process.WorkItem;
 import org.kie.kogito.process.bpmn2.BpmnProcess;
 import org.kie.kogito.process.bpmn2.BpmnProcessInstance;
@@ -32,18 +34,38 @@ import org.kie.kogito.process.bpmn2.BpmnVariables;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.kie.kogito.internal.process.runtime.KogitoProcessInstance.STATE_ACTIVE;
 import static org.kie.kogito.internal.process.runtime.KogitoProcessInstance.STATE_COMPLETED;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-class JdbcProcessInstancesIT extends TestHelper {
+abstract class AbstractProcessInstancesIT {
 
-    @ParameterizedTest
-    @MethodSource("datasources")
-    void testBasicTaskFlow(DataSource dataSource) {
-        var factory = new TestProcessInstancesFactory(dataSource, false);
+    public static final String TEST_ID = "02ac3854-46ee-42b7-8b63-5186c9889d96";
+    public static SecurityPolicy securityPolicy = SecurityPolicy.of(IdentityProviders.of("john"));
+
+    public static BpmnProcess createProcess(TestProcessInstancesFactory factory, String fileName) {
+        BpmnProcess process = BpmnProcess.from(new ClassPathResource(fileName)).get(0);
+        process.setProcessInstancesFactory(factory);
+        process.configure();
+        process.instances().values(ProcessInstanceReadMode.MUTABLE).forEach(p -> p.abort());
+        return process;
+    }
+
+    public static BpmnProcess configure(boolean lock) {
+        BpmnProcess process = BpmnProcess.from(new ClassPathResource("BPMN2-UserTask-Script.bpmn2")).get(0);
+        process.setProcessInstancesFactory(new TestProcessInstancesFactory(null, lock));
+        process.configure();
+        return process;
+    }
+
+    abstract DataSource getDataSource();
+
+    @Test
+    void testBasicTaskFlow() {
+        var factory = new TestProcessInstancesFactory(getDataSource(), false);
         BpmnProcess process = createProcess(factory, "BPMN2-UserTask.bpmn2");
         ProcessInstance<BpmnVariables> processInstance = process.createInstance(BpmnVariables.create(Collections
                 .singletonMap("test", "test")));
@@ -76,10 +98,9 @@ class JdbcProcessInstancesIT extends TestHelper {
         assertThat(process.instances().values()).isEmpty();
     }
 
-    @ParameterizedTest
-    @MethodSource("datasources")
-    void testBasicFlow(DataSource dataSource) {
-        var factory = new TestProcessInstancesFactory(dataSource, false);
+    @Test
+    void testBasicFlow() {
+        var factory = new TestProcessInstancesFactory(getDataSource(), false);
         BpmnProcess process = createProcess(factory, "BPMN2-UserTask.bpmn2");
         ProcessInstance<BpmnVariables> processInstance = process.createInstance(BpmnVariables.create(Collections
                 .singletonMap("test",
@@ -108,5 +129,61 @@ class JdbcProcessInstancesIT extends TestHelper {
         JDBCProcessInstances processInstances = (JDBCProcessInstances) process.instances();
         assertThatExceptionOfType(RuntimeException.class).isThrownBy(() -> processInstances.findById(TEST_ID));
         assertThatExceptionOfType(RuntimeException.class).isThrownBy(() -> processInstances.remove(TEST_ID));
+    }
+
+    @Test
+    public void testUpdate() {
+        var factory = new TestProcessInstancesFactory(getDataSource(), true);
+        BpmnProcess process = createProcess(factory, "BPMN2-UserTask.bpmn2");
+        ProcessInstance<BpmnVariables> processInstance = process.createInstance(BpmnVariables.create(Collections.singletonMap("test", "test")));
+        processInstance.start();
+
+        JDBCProcessInstances processInstances = (JDBCProcessInstances) process.instances();
+        assertThat(processInstances.size()).isOne();
+        Optional<?> foundOne = processInstances.findById(processInstance.id());
+        BpmnProcessInstance instanceOne = (BpmnProcessInstance) foundOne.get();
+        foundOne = processInstances.findById(processInstance.id());
+        BpmnProcessInstance instanceTwo = (BpmnProcessInstance) foundOne.get();
+        assertEquals(1L, instanceOne.version());
+        assertEquals(1L, instanceTwo.version());
+        instanceOne.updateVariables(BpmnVariables.create(Collections.singletonMap("s", "test")));
+        try {
+            BpmnVariables testvar = BpmnVariables.create(Collections.singletonMap("ss", "test"));
+            instanceTwo.updateVariables(testvar);
+        } catch (RuntimeException e) {
+            assertThat(e.getMessage()).isEqualTo("The document with ID: " + instanceOne.id() + " was updated or deleted by other request.");
+        }
+        foundOne = processInstances.findById(processInstance.id());
+        instanceOne = (BpmnProcessInstance) foundOne.get();
+        assertEquals(2L, instanceOne.version());
+
+        processInstances.remove(processInstance.id());
+        assertThat(processInstances.size()).isZero();
+        assertThat(process.instances().values()).isEmpty();
+    }
+
+    @Test
+    public void testRemove() {
+        var factory = new TestProcessInstancesFactory(getDataSource(), true);
+        BpmnProcess process = createProcess(factory, "BPMN2-UserTask.bpmn2");
+        ProcessInstance<BpmnVariables> processInstance = process.createInstance(BpmnVariables.create(Collections.singletonMap("test", "test")));
+        processInstance.start();
+
+        JDBCProcessInstances processInstances = (JDBCProcessInstances) process.instances();
+        assertThat(processInstances.size()).isOne();
+        Optional<?> foundOne = processInstances.findById(processInstance.id());
+        BpmnProcessInstance instanceOne = (BpmnProcessInstance) foundOne.get();
+        foundOne = processInstances.findById(processInstance.id());
+        BpmnProcessInstance instanceTwo = (BpmnProcessInstance) foundOne.get();
+        assertEquals(1L, instanceOne.version());
+        assertEquals(1L, instanceTwo.version());
+
+        processInstances.remove(instanceOne.id());
+        try {
+            String id = instanceTwo.id();
+            processInstances.remove(id);
+        } catch (RuntimeException e) {
+            assertThat(e.getMessage()).isEqualTo("The document with ID: " + instanceOne.id() + " was deleted by other request.");
+        }
     }
 }
