+++ integration-tests/integration-tests-quarkus-norest/pom.xml
@@ -45,10 +45,6 @@
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-resteasy-jackson</artifactId>
     </dependency>
-    <dependency>
-      <groupId>io.quarkus</groupId>
-      <artifactId>quarkus-smallrye-reactive-messaging-kafka</artifactId>
-    </dependency>
 
     <dependency>
       <groupId>io.quarkus</groupId>
