+++ integration-tests/integration-tests-quarkus-processes-persistence/integration-tests-quarkus-processes-jdbc/src/main/resources/application.properties
@@ -2,3 +2,4 @@
 quarkus.http.test-port=0
 
 kogito.persistence.type=jdbc
+quarkus.datasource.db-kind=postgresql
