+++ integration-tests/integration-tests-quarkus-processes-persistence/integration-tests-quarkus-processes-jdbc/src/test/resources/application.properties
@@ -1,6 +1,6 @@
 # Quarkus
 quarkus.http.test-port=0
-
+quarkus.datasource.devservices.enabled=false
 
 kogito.persistence.type=postgresql
