+++ integration-tests/integration-tests-quarkus-processes-persistence/integration-tests-quarkus-processes-mongodb/src/test/resources/application.properties
@@ -4,3 +4,4 @@ quarkus.http.test-port=0
 kogito.persistence.type=mongodb
 # MongoDB
 quarkus.mongodb.database=kogito
+quarkus.mongodb.devservices.enabled=false
