+++ kogito-build/kogito-dependencies-bom/pom.xml
@@ -16,7 +16,7 @@
 
   <properties>
     <!-- this version property is used in plugins but also in dependencies too -->
-    <version.io.quarkus>2.4.0.Final</version.io.quarkus>
+    <version.io.quarkus>2.4.2.Final</version.io.quarkus>
     <version.io.quarkus.quarkus-test-maven>2.4.0.Final</version.io.quarkus.quarkus-test-maven>
     <version.org.springframework.boot>2.3.10.RELEASE</version.org.springframework.boot>
 
@@ -72,7 +72,7 @@
     <version.org.keycloak>14.0.0</version.org.keycloak>
     <version.org.confluent.kafka>5.4.3</version.org.confluent.kafka>
     <version.org.mongo>4.3.3</version.org.mongo>
-    <version.org.mongo-image>4.2.3</version.org.mongo-image> <!-- Version 4.3.0 is not existing ... -->
+    <version.org.mongo-image>4.4.10</version.org.mongo-image> <!-- Version 4.3.0 is not existing ... -->
     <version.org.mongo.springboot>4.1.0</version.org.mongo.springboot> <!-- https://issues.redhat.com/browse/KOGITO-5031 -->
     <version.org.mozilla.rhino>1.7.13</version.org.mozilla.rhino>
 
@@ -86,7 +86,7 @@
     <version.org.junit.platform>1.8.1</version.org.junit.platform> <!-- otherwise Quarkus brings its own, silently disabling some tests -->
     <version.org.junit.pioneer>1.4.2</version.org.junit.pioneer>
     <version.org.mockito>3.12.4</version.org.mockito>
-    <version.org.testcontainers>1.15.3</version.org.testcontainers>
+    <version.org.testcontainers>1.16.2</version.org.testcontainers>
 
     <version.org.mvel>2.4.12.Final</version.org.mvel>
     <version.org.reactivestreams>1.0.3</version.org.reactivestreams>
