+++ kogito-test-utils/pom.xml
@@ -72,6 +72,7 @@
       <groupId>org.testcontainers</groupId>
       <artifactId>jdbc</artifactId>
       <version>${version.org.testcontainers}</version>
+      <scope>compile</scope>
     </dependency>
 
     <dependency>
