+++ quarkus/addons/mail/runtime/src/test/resources/application.properties
@@ -1,3 +1,1 @@
-# Quarkus
-quarkus.http.test-port=0
 quarkus.kafka.devservices.enabled=false
